+++ src/main/java/essentialclient/mixins/soulSpeedFovMultiplier/AbstractClientPlayerEntityMixin.java
@@ -1,4 +1,4 @@
-package essentialclient.mixins.soulBlocksEffectOnFov;
+package essentialclient.mixins.soulSpeedFovMultiplier;
 
 import essentialclient.feature.clientrule.ClientRules;
 import net.minecraft.client.network.AbstractClientPlayerEntity;
